refactor: 코드 최적화 - 변경된 명칭 적용 - import 정리 - reformat code
