+++ src/main/java/wordle/Application.java
@@ -1,6 +1,6 @@
 package wordle;
 
-public class Main {
+public class Application {
     public static void main(String[] args) {
         GameMachine gameMachine = new GameMachine();
         gameMachine.start();
