+++ src/main/java/wordle/domain/Letter.java
@@ -14,7 +14,7 @@ public class Letter {
 
     private void validateAlpha(char ch) {
         // Letter 라는 클래스가 하나의 글자
-        if (ch > 'z' || ch <'a') {
+        if (ch > 'z' || ch < 'a') {
             throw new IllegalArgumentException("shoud be alphabet");
         }
     }
