+++ src/main/java/wordle/domain/Round.java
@@ -2,7 +2,7 @@ package wordle.domain;
 
 public class Round {
 
-    private static final int MAX_STAGE = 6;
+    private static final int MAX_ROUND = 6;
     private int current = 1;
 
     public void next() {
@@ -10,11 +10,11 @@ public class Round {
     }
 
     public boolean isFinal() {
-        return this.current >= MAX_STAGE;
+        return this.current >= MAX_ROUND;
     }
 
     @Override
     public String toString() {
-        return current + "/" + MAX_STAGE;
+        return current + "/" + MAX_ROUND;
     }
 }
