+++ src/test/java/wordle/util/FileReaderTest.java
@@ -1,8 +1,7 @@
-package wordle;
+package wordle.util;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import wordle.util.FileReader;
 
 import java.util.List;
